fix(ignis)!: remove system tray to reduce cpu usage
It seems that the system tray has high cpu usage if a system tray item
updates its icon very frequently. This, for example, seems to happen for
the nextcloud system tray. For now, we deactive the system tray,
completely until I can come up with a fix for this issue.